# BankingSystemAPIs/apidocs.json
{
    "openapi": "3.0.1",
    "info": {
        "title": "Banking System API Documentation",
        "description": "The Banking System API provides a comprehensive set of endpoints for managing banking-related operations. It is designed to support various banking activities such as account management, transaction processing, and exchange rate management. This API is built with scalability and security in mind, ensuring seamless integration with client applications.\r\n\r\nKey Features:\r\nAccount Management: Create, retrieve, update, and delete accounts. Perform operations such as deposits, withdrawals, and transfers.\r\nTransaction Processing: Retrieve transaction histories, manage transaction records, and delete transactions securely.\r\nExchange Rates: Manage exchange rates for supported currencies, including retrieval, creation, updates, and deletions.\r\nLogging: Detailed request and response logging to facilitate debugging and auditing.\r\nRESTful Design: Follows RESTful principles with predictable endpoints, standard HTTP methods, and meaningful status codes.\r\nSwagger Documentation: Integrated Swagger/OpenAPI support for testing and visualizing API capabilities.",
        "contact": {
            "name": "Mahmoud Metwalli",
            "email": "mammkbih@gmail.com"
        },
        "version": "v1"
    },
    "paths": {
        "/api/account/savings": {
            "post": {
                "tags": [
                    "Account"
                ],
                "summary": "Creates a new savings account.",
                "requestBody": {
                    "description": "Details of the savings account.",
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/SavingsAccountRequest"
                            }
                        },
                        "text/json": {
                            "schema": {
                                "$ref": "#/components/schemas/SavingsAccountRequest"
                            }
                        },
                        "application/*+json": {
                            "schema": {
                                "$ref": "#/components/schemas/SavingsAccountRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "text/plain": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProblemDetails"
                                }
                            },
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProblemDetails"
                                }
                            },
                            "text/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProblemDetails"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/account/checking": {
            "post": {
                "tags": [
                    "Account"
                ],
                "summary": "Creates a new checking account.",
                "requestBody": {
                    "description": "Details of the checking account.",
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CheckingAccountRequest"
                            }
                        },
                        "text/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CheckingAccountRequest"
                            }
                        },
                        "application/*+json": {
                            "schema": {
                                "$ref": "#/components/schemas/CheckingAccountRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "text/plain": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProblemDetails"
                                }
                            },
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProblemDetails"
                                }
                            },
                            "text/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProblemDetails"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/account/{id}": {
            "put": {
                "tags": [
                    "Account"
                ],
                "summary": "Updates an account by ID.",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Account ID.",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Updated account details.",
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Account"
                            }
                        },
                        "text/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Account"
                            }
                        },
                        "application/*+json": {
                            "schema": {
                                "$ref": "#/components/schemas/Account"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "text/plain": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProblemDetails"
                                }
                            },
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProblemDetails"
                                }
                            },
                            "text/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProblemDetails"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "text/plain": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProblemDetails"
                                }
                            },
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProblemDetails"
                                }
                            },
                            "text/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProblemDetails"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Account"
                ],
                "summary": "Deletes an account by ID.",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Account ID.",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "text/plain": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProblemDetails"
                                }
                            },
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProblemDetails"
                                }
                            },
                            "text/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProblemDetails"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/account/{id}/transactions": {
            "get": {
                "tags": [
                    "Account"
                ],
                "summary": "Retrieves all transactions for an account.",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Account ID.",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "text/plain": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProblemDetails"
                                }
                            },
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProblemDetails"
                                }
                            },
                            "text/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProblemDetails"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/account/{id}/transactions/source": {
            "get": {
                "tags": [
                    "Account"
                ],
                "summary": "Retrieves source transactions for an account.",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Account ID.",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "text/plain": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProblemDetails"
                                }
                            },
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProblemDetails"
                                }
                            },
                            "text/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProblemDetails"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/account/{id}/transactions/target": {
            "get": {
                "tags": [
                    "Account"
                ],
                "summary": "Retrieves target transactions for an account.",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Account ID.",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "text/plain": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProblemDetails"
                                }
                            },
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProblemDetails"
                                }
                            },
                            "text/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProblemDetails"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Account": {
                "required": [
                    "accountNumber",
                    "balance",
                    "clientId",
                    "currency"
                ],
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "accountNumber": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "balance": {
                        "type": "number",
                        "format": "double"
                    },
                    "rowVersion": {
                        "type": "string",
                        "format": "byte",
                        "nullable": true
                    },
                    "currency": {
                        "maxLength": 3,
                        "minLength": 1,
                        "type": "string"
                    },
                    "rate": {
                        "$ref": "#/components/schemas/Rate"
                    },
                    "clientId": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "accountsTransactions": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/AccountTransaction"
                        },
                        "nullable": true,
                        "readOnly": true
                    },
                    "client": {
                        "$ref": "#/components/schemas/Client"
                    }
                },
                "additionalProperties": false
            },
            "AccountTransaction": {
                "type": "object",
                "properties": {
                    "accountId": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "transactionId": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "account": {
                        "$ref": "#/components/schemas/Account"
                    },
                    "transaction": {
                        "$ref": "#/components/schemas/Transaction"
                    },
                    "source": {
                        "type": "boolean"
                    }
                },
                "additionalProperties": false
            },
            "CheckingAccountRequest": {
                "required": [
                    "accountNumber",
                    "balance",
                    "currency",
                    "overDraft",
                    "userId"
                ],
                "type": "object",
                "properties": {
                    "userId": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "currency": {
                        "maxLength": 3,
                        "minLength": 1,
                        "type": "string"
                    },
                    "accountNumber": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "balance": {
                        "minimum": 0.01,
                        "type": "number",
                        "format": "double"
                    },
                    "overDraft": {
                        "minimum": 0.01,
                        "type": "number",
                        "format": "double"
                    }
                },
                "additionalProperties": false
            },
            "Client": {
                "required": [
                    "email",
                    "name"
                ],
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "name": {
                        "maxLength": 100,
                        "minLength": 1,
                        "type": "string"
                    },
                    "email": {
                        "minLength": 1,
                        "type": "string",
                        "format": "email"
                    },
                    "accounts": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Account"
                        },
                        "nullable": true
                    }
                },
                "additionalProperties": false
            },
...
            "ProblemDetails": {
                "type": "object",
                "properties": {
                    "type": {
                        "type": "string",
                        "nullable": true
                    },
                    "title": {
                        "type": "string",
                        "nullable": true
                    },
                    "status": {
                        "type": "integer",
                        "format": "int32",
                        "nullable": true
                    },
                    "detail": {
                        "type": "string",
                        "nullable": true
                    },
                    "instance": {
                        "type": "string",
                        "nullable": true
                    }
                },
                "additionalProperties": {}
            },
            "Rate": {
                "required": [
                    "currency"
                ],
                "type": "object",
                "properties": {
                    "currency": {
                        "maxLength": 3,
                        "minLength": 1,
                        "type": "string"
                    },
                    "currencyRate": {
                        "type": "number",
                        "format": "double"
                    }
                },
                "additionalProperties": false
            },
            "SavingsAccountRequest": {
                "required": [
                    "accountNumber",
                    "balance",
                    "currency",
                    "interest",
                    "userId"
                ],
                "type": "object",
                "properties": {
                    "userId": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "currency": {
                        "maxLength": 3,
                        "minLength": 1,
                        "type": "string"
                    },
                    "accountNumber": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "balance": {
                        "minimum": 0.01,
                        "type": "number",
                        "format": "double"
                    },
                    "interest": {
                        "maximum": 100,
                        "minimum": 0.01,
                        "type": "number",
                        "format": "double"
                    }
                },
                "additionalProperties": false
            },
            "Transaction": {
                "required": [
                    "currency"
                ],
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "accountsTransactions": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/AccountTransaction"
                        },
                        "nullable": true
                    },
                    "amount": {
                        "type": "number",
                        "format": "double"
                    },
                    "currency": {
                        "maxLength": 3,
                        "minLength": 1,
                        "type": "string"
                    },
                    "rate": {
                        "$ref": "#/components/schemas/Rate"
                    },
                    "timestamp": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "additionalProperties": false
            },
...
        }
    }
}